Isolate --new/--last test for test stability
